+++ tests/test_pypssalib.py
@@ -4,7 +4,7 @@ import pytest
 
 
 def test_version():
-    assert m.__version__ == "0.1.0-dev3"
+    assert m.__version__ == "0.1.0-dev4"
 
 
 def test_method():